@@ -5,9 +5,9 @@
   "dependencies": {
     "bootstrap": "^4.3.1",
     "prop-types": "^15.7.2",
-    "react": "^16.8.6",
+    "react": "^16.10.2",
     "react-bootstrap": "^1.0.0-beta.10",
-    "react-dom": "^16.8.6",
+    "react-dom": "^16.10.2",
     "react-router-dom": "^5.0.1",
     "react-scripts": "3.0.1",
     "reactstrap": "^8.0.1"
