@@ -8,6 +8,7 @@ export default class FoodEditForm extends Component {
     carb: 0,
     fat: 0,
     protein: 0,
+    date: 0,
     count: 0
   };
 
@@ -19,15 +20,18 @@ export default class FoodEditForm extends Component {
 
   updateExistingFood = evt => {
     evt.preventDefault();
+    let d = new Date();
+
 
     const editedFood = {
       id: this.props.match.params.foodId,
       userId: parseInt(sessionStorage.getItem("userId")),
       name: this.state.name,
-      fat: this.state.fat,
-      carb: this.state.carb,
-      protein: this.state.protein,
-      count: this.state.count
+      fat: Number(this.state.fat),
+      carb: Number(this.state.carb),
+      protein: Number(this.state.protein),
+      count: this.state.count,
+      date: d.getDate()
     };
     DataManager.get("foods", this.props.match.params.foodId)
       .then(food => {
